Added light support, need to change the way I update the lights ubo

+++ Engine/se_swap_chain.hpp
@@ -27,7 +27,7 @@ namespace se
     SESwapChain(const SESwapChain &) = delete;
     SESwapChain &operator=(const SESwapChain &) = delete;
 
-	int getCurrentFrame() const { return currentFrame; }
+	size_t getCurrentFrame() const { return currentFrame; }
     VkFramebuffer getFrameBuffer(int index) { return swapChainFramebuffers[index]; }
     VkRenderPass getRenderPass() { return renderPass; }
     VkImageView getImageView(int index) { return swapChainImageViews[index]; }
